Removing gallery screen and integration with CameraViewModel
