+++ app/src/main/java/com/android/feedme/ui/camera/Camera.kt
@@ -7,6 +7,8 @@ import android.content.pm.PackageManager
 import android.graphics.Bitmap
 import android.graphics.Matrix
 import android.util.Log
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.camera.core.ImageCapture
 import androidx.camera.core.ImageCaptureException
 import androidx.camera.core.ImageProxy
@@ -66,7 +68,6 @@ import com.android.feedme.ui.navigation.NavigationActions
 import com.android.feedme.ui.navigation.TopBarNavigation
 import com.android.feedme.ui.theme.BottomIconColorSelected
 import com.android.feedme.ui.theme.CameraButtonsBackground
-import kotlinx.coroutines.launch
 
 /**
  * A composable function representing the camera screen.
@@ -90,6 +91,7 @@ fun CameraScreen(navigationActions: NavigationActions, inputViewModel: InputView
 
   ///// Machine Learning Part /////
 
+
   val applicationContext = LocalContext.current
   // Request camera permission if not already granted
   if (!hasRequiredPermissions(applicationContext)) {
@@ -108,8 +110,9 @@ fun CameraScreen(navigationActions: NavigationActions, inputViewModel: InputView
   val cameraViewModel = viewModel<CameraViewModel>()
   val bitmaps by cameraViewModel.bitmaps.collectAsState()
   val photoSavedMessageVisible by cameraViewModel.photoSavedMessageVisible.collectAsState()
+    val pickImage = cameraViewModel.galleryLauncher()
 
-  val listOfIngredientToInput = cameraViewModel.listOfIngredientToInput.collectAsState()
+    val listOfIngredientToInput = cameraViewModel.listOfIngredientToInput.collectAsState()
 
   BottomSheetScaffold(
       modifier = Modifier.testTag("CameraScreen"),
@@ -144,7 +147,11 @@ fun CameraScreen(navigationActions: NavigationActions, inputViewModel: InputView
                             .padding(10.dp)
                             .testTag("GalleryButton"),
                     // Open the local gallery when the gallery button is clicked
-                    onClick = { scope.launch { scaffoldState.bottomSheetState.expand() } }) {
+                    onClick = {
+                        pickImage.launch(
+                        PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
+                    )
+                    }) {
                       Icon(imageVector = Icons.Default.Photo, contentDescription = "Open gallery")
                     }
